+++ packages/ui5-app/webapp/controller/Main.controller.js
@@ -1,7 +1,7 @@
 sap.ui.define([
     "test/Sample/controller/BaseController",
     "sap/m/MessageToast"
-], (Controller, MessageToast) => {
+], (Controller, MessageToast, Chart) => {
     "use strict";
 
     return Controller.extend("test.Sample.controller.Main", {