+++ packages/ui5-tooling-modules/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.2.1...ui5-tooling-modules@0.2.2) (2022-03-04)
+
+
+### Bug Fixes
+
+* **ui5-tooling-modules:** ensure to export "exports" not default for cjs ([5c39eae](https://github.com/ui5-community/ui5-ecosystem-showcase/commit/5c39eae5a0db2ff8ad46fcfdec804c13da0cfd32)), closes [#589](https://github.com/ui5-community/ui5-ecosystem-showcase/issues/589)
+
+
+
+
+
+## [0.2.1](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.2.0...ui5-tooling-modules@0.2.1) (2022-02-27)
+
+
+### Bug Fixes
+
+* **ui5-tooling-modules:** skip css and simulate process.env ([#585](https://github.com/ui5-community/ui5-ecosystem-showcase/issues/585)) ([e6f246c](https://github.com/ui5-community/ui5-ecosystem-showcase/commit/e6f246c057228c9b3ae3c45fcc952fbcd8375a96))
+
+
+
+
+
+# [0.2.0](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.1.3...ui5-tooling-modules@0.2.0) (2022-02-24)
+
+
+### Features
+
+* **ui5-tooling-modules:** extended UI5 modules support ([#583](https://github.com/ui5-community/ui5-ecosystem-showcase/issues/583)) ([fe2c8a7](https://github.com/ui5-community/ui5-ecosystem-showcase/commit/fe2c8a72c9fa5e906db56ecf86c83621d0585eb8)), closes [#578](https://github.com/ui5-community/ui5-ecosystem-showcase/issues/578)
+
+
+
+
+
+## [0.1.3](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.1.2...ui5-tooling-modules@0.1.3) (2022-02-22)
+
+**Note:** Version bump only for package ui5-tooling-modules
+
+
+
+
+
 ## [0.1.2](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.1.1...ui5-tooling-modules@0.1.2) (2022-02-07)